# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Blockchain API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/blocks/{networkCode}/{hash}": {
            "get": {
                "description": "Get a block along with the first ten transactions",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Blocks"
                ],
                "summary": "Get a block",
                "operationId": "get-block",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The acronym of the network you're querying required",
                        "name": "networkCode",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "The blockhash or height (number) on the network you're querying",
                        "name": "hash",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/presenter.BlockResponse"
                        }
                    }
                }
            }
        },
        "/transactions/{networkCode}/{transactionId}": {
            "get": {
                "description": "Get a transaction",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transactions"
                ],
                "summary": "Get a transaction",
                "operationId": "get-transaction",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The acronym of the network you're querying required",
                        "name": "networkCode",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "The transaction hash (id) on the network you're querying",
                        "name": "transactionId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/presenter.TransactionResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "presenter.BlockResponse": {
            "type": "object",
            "properties": {
                "blockNumber": {
                    "description": "The height of the block in the blockchain, or its number",
                    "type": "integer"
                },
                "dateTime": {
                    "description": "The time at which this block was mined by the miner",
                    "type": "string"
                },
                "networkCode": {
                    "description": "The acronym of the network",
                    "type": "string"
                },
                "nextBlockhash": {
                    "description": "The block hash of the next block in the blockchain. NextBlockhash=null if this is the last block in the blockchain",
                    "type": "string"
                },
                "previousBlockhash": {
                    "description": "The block hash of the previous block in the blockchain",
                    "type": "string"
                },
                "size": {
                    "description": "The size of the block in bytes",
                    "type": "integer"
                },
                "transactions": {
                    "description": "The array of ids of all transactions in this block, starting with the newly generated coins (only the first 10)",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/presenter.TransactionResponse"
                    }
                }
            }
        },
        "presenter.TransactionResponse": {
            "type": "object",
            "properties": {
                "dateTime": {
                    "description": "The time at which this transaction received by SoChain, or was mined by the miner",
                    "type": "string"
                },
                "fee": {
                    "description": "The fee paid to the miner",
                    "type": "number"
                },
                "sentValue": {
                    "description": "The total value of all coins sent in this transaction",
                    "type": "number"
                },
                "transactionId": {
                    "description": "The transaction id",
                    "type": "string"
                }
            }
        }
    }
}